requirements.txt
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.3.56
Pillow==8.3.1
transformers==4.10.0
numpy==1.21.1
scikit-learn==0.24.2
matplotlib==3.4.2
pandas==1.3.1